refactor Camera component
