+++ client/features/camera/CamBtn.jsx
@@ -1,0 +1,23 @@
+import React from 'react'
+import { useDispatch } from 'react-redux'
+import { useHistory } from 'react-router-dom'
+
+import { setImg } from './camProtoSlice'
+
+function CamBtn () {
+  const dispatch = useDispatch()
+  const history = useHistory()
+
+  function addImg (e) {
+    const img = e.target.files[0]
+    const imgURL = URL.createObjectURL(img)
+    dispatch(setImg(imgURL))
+    history.push('/caption')
+  }
+
+  return (
+    <input type='file' id='imageFile' capture='environment' accept='image/*' onChange={addImg}/>
+  )
+}
+
+export default CamBtn
